requirements.txt
Django==1.11.2
django-solo==1.1.2
django-wysiwyg-redactor==0.5.1
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.1
pytz==2017.2
